… CMakeLists.txt
… examples/llama-bench/llama-bench.cpp
… ggml-backend.h
… ggml-blas.cpp
… ggml-cuda.cu
… ggml-kompute.cpp
… ggml-metal.m
… ggml-rpc.cpp
… ggml-sycl.cpp
… ggml-vulkan.cpp
… ggml.c
… llama.cpp