Moves BLAS support from `ggml.c` to a separate backend, and adds the necessary changes to ggml-backend to support backends that only implement matrix multiplication.

- Changes to ggml-backend
    - Support for fallback to CPU with `ggml_backend_sched`
        - Operations not implemented by a backend will be automatically run on the CPU, as long as the operation is reported as not supported in the `supports_op` function of the backend
    - Moved buffer type function `supports_backend` to backend function `supports_buft`
    - Backends that want to declare compatibility with any kind of host buffer can return `ggml_backend_buft_is_host` from `supports_buft`
    - `ggml_backend_sched` will avoid copies between backends when the backend supports the buffer type
        - Eg. when switching from Metal to CPU, no tensors will be copied since the Metal buffers are compatible with the CPU backend (but not the other way around)
- The `GGML_SCHED_DEBUG` environment variable can be used to view the graph splits. This is useful to see what operations are being run on each backend
- Adds the BLAS backend
    - Supports matrix multiplication using a BLAS library. Previously, this was supported as part of the CPU backend
    - Threads are no longer spinning while BLAS is running, potentially improving performance, and batch processing is no longer limited to 4 threads when using BLAS
    - The number of threads of the BLAS library configured automatically for OpenBLAS and BLIS (with `-t` or `-tb`)
    - For better performance, it is recommended to use OpenMP versions of the BLAS libraries, if available (except macOS)
    - Like before, to enable the BLAS backend, build with the flag `LLAMA_BLAS` when using cmake, or when using make, `LLAMA_OPENBLAS`, `LLAMA_OPENBLAS64` or `LLAMA_BLIS`
    - On macOS, this is enabled by default through Accelerate
    - BLAS support has been removed from the CPU backend in `ggml.c`. Applications that want to support BLAS will need to use the BLAS backend
    - Since this backend only implements matrix multiplication, it should be used with `ggml_backend_sched` alongside the CPU or other backends to provide support for other operations
    - Note: the BLAS backend should not be used alongside GPU backends, as it will prevent offloading of large batches with partial offloading