@ggerganov 2024-03-21
> this will also have the effect that using BLAS will require using ggml-backend and ggml_backend_sched, is that a problem?

Will just need to adapt `whisper.cpp` when it's ready