@zhouwg 2024-05-30
> > this will also have the effect that using BLAS will require using ggml-backend and ggml_backend_sched, is that a problem?
> 
> Will just need to adapt `whisper.cpp` when it's ready

I had been submitted a PR to address this and it works very well with whisper.cpp and llama.cpp using QNN backend.