@slaren 2024-05-30
@mofosyne I appreciate that you are trying to help, but please don't do that on my PRs. I very often have uncommitted local changes and I prefer to deal with the merge conflicts myself.