@slaren 2024-06-06
@ggerganov I am thinking about how accelerate should interact with the BLAS backend. I think this would make sense:
- `GGML_USE_ACCELERATE` defined: accelerate is used in `ggml.c`
- `GGML_USE_ACCELERATE` and `GGML_USE_BLAS` defined: the BLAS backend is used using accelerate as the BLAS library

Conversely:
- If `GGML_USE_BLAS` is not defined, accelerate will not be used for GEMMs

Currently `llama.cpp` has to check check for `defined(GGML_USE_BLAS) || defined(GGML_USE_ACCELERATE)` to decide when to use the BLAS backend, which doesn't seem very good. In the BLAS backend, accelerate is treated in the same way as any other BLAS library.