@ggerganov 2024-06-06
Yes, that makes sense. With only `GGML_USE_ACCELERATE` we will still be able to use some non-GEMM functionality from the Accelerate framework such as `vDSP` - it's just very convenient for Apple Silicon devices to have this framework available in the core `ggml.c`. For GEMMs we would explicitly need to have `GGML_USE_BLAS` for all kinds of BLAS implementations, including Accelerate's BLAS