@slaren 2024-06-07
Some notes on performance.

BLIS responds very well to these changes, and improves performance in most cases. In master, the number of threads is limited to 4 when using BLAS. This is not the case with the BLAS backend. Using a higher number of threads should improve the performance of dequantizing and other operations.


`LLAMA_BLIS=1 BLIS_NUM_THREADS=24`:

| CPU       | Model            |   Model Size [GiB] |   Threads | Test   |   t/s master |   t/s sl/blas-backend |   Speedup |
|:----------|:-----------------|-------------------:|----------:|:-------|-------------:|----------------------:|----------:|
| i9-13900K | llama 7B Q4_0    |               3.56 |         1 | pp32   |        10.03 |                 11.13 |      1.11 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         1 | pp64   |        17.65 |                 19.41 |      1.10 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         1 | pp128  |        29.66 |                 28.09 |      0.95 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         4 | pp32   |        15.58 |                 20.81 |      1.34 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         4 | pp64   |        26.79 |                 32.48 |      1.21 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         4 | pp128  |        39.35 |                 46.65 |      1.19 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         8 | pp32   |        15.79 |                 22.86 |      1.45 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         8 | pp64   |        27.02 |                 35.84 |      1.33 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         8 | pp128  |        39.06 |                 48.96 |      1.25 |
| i9-13900K | llama 7B Q4_0    |               3.56 |        16 | pp32   |        15.37 |                 22.30 |      1.45 |
| i9-13900K | llama 7B Q4_0    |               3.56 |        16 | pp64   |        26.78 |                 28.70 |      1.07 |
| i9-13900K | llama 7B Q4_0    |               3.56 |        16 | pp128  |        38.53 |                 41.96 |      1.09 |
| i9-13900K | llama 7B all F32 |              25.10 |         1 | pp32   |        30.91 |                 49.17 |      1.59 |
| i9-13900K | llama 7B all F32 |              25.10 |         1 | pp64   |        43.69 |                 61.69 |      1.41 |
| i9-13900K | llama 7B all F32 |              25.10 |         1 | pp128  |        56.80 |                 66.45 |      1.17 |
| i9-13900K | llama 7B all F32 |              25.10 |         4 | pp32   |        29.73 |                 51.31 |      1.73 |
| i9-13900K | llama 7B all F32 |              25.10 |         4 | pp64   |        44.25 |                 64.98 |      1.47 |
| i9-13900K | llama 7B all F32 |              25.10 |         4 | pp128  |        58.12 |                 74.17 |      1.28 |
| i9-13900K | llama 7B all F32 |              25.10 |         8 | pp32   |        28.94 |                 49.29 |      1.70 |
| i9-13900K | llama 7B all F32 |              25.10 |         8 | pp64   |        45.10 |                 68.30 |      1.51 |
| i9-13900K | llama 7B all F32 |              25.10 |         8 | pp128  |        58.40 |                 78.17 |      1.34 |
| i9-13900K | llama 7B all F32 |              25.10 |        16 | pp32   |        29.68 |                 32.29 |      1.09 |
| i9-13900K | llama 7B all F32 |              25.10 |        16 | pp64   |        43.85 |                 44.14 |      1.01 |
| i9-13900K | llama 7B all F32 |              25.10 |        16 | pp128  |        59.05 |                 58.07 |      0.98 |


OpenBLAS is a headache. Increasing the number of threads above 8 has catastrophic effects on performance in ways that I cannot explain, because these threads are not running while the OpenBLAS gemm is running. Since there are better alternatives easily available now, I would suggest ignoring this library.

`LLAMA_OPENBLAS=1 OPENBLAS_NUM_THREADS=24`:

| CPU       | Model            |   Model Size [GiB] |   Threads | Test   |   t/s master |   t/s sl/blas-backend |   Speedup |
|:----------|:-----------------|-------------------:|----------:|:-------|-------------:|----------------------:|----------:|
| i9-13900K | llama 7B Q4_0    |               3.56 |         4 | pp32   |        14.79 |                 14.63 |      0.99 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         4 | pp64   |        22.30 |                 21.85 |      0.98 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         4 | pp128  |        29.36 |                 28.44 |      0.97 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         8 | pp32   |        15.01 |                 15.83 |      1.05 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         8 | pp64   |        22.22 |                 23.77 |      1.07 |
| i9-13900K | llama 7B Q4_0    |               3.56 |         8 | pp128  |        29.38 |                 29.97 |      1.02 |
| i9-13900K | llama 7B all F32 |              25.10 |         4 | pp32   |        23.55 |                 23.55 |      1.00 |
| i9-13900K | llama 7B all F32 |              25.10 |         4 | pp64   |        30.88 |                 30.46 |      0.99 |
| i9-13900K | llama 7B all F32 |              25.10 |         4 | pp128  |        36.20 |                 34.92 |      0.96 |
| i9-13900K | llama 7B all F32 |              25.10 |         8 | pp32   |        23.60 |                 22.81 |      0.97 |
| i9-13900K | llama 7B all F32 |              25.10 |         8 | pp64   |        30.84 |                 31.13 |      1.01 |
| i9-13900K | llama 7B all F32 |              25.10 |         8 | pp128  |        34.42 |                 35.92 |      1.04 |


The OpenMP thread pool does not work on macOS, so the overhead of starting threads continuously is expected to be higher. However, with the default number of threads on my system (12), it still seems to result in an speedup. There is a drop in performance specifically with 4 threads and the larger batch sizes that I don't understand, I would expect the overhead to be proportionally smaller as the batch size increases.

| CPU   | Model            |   Model Size [GiB] |   Threads | Test   |   t/s master |   t/s sl/blas-backend |   Speedup |
|:------|:-----------------|-------------------:|----------:|:-------|-------------:|----------------------:|----------:|
|       | llama 7B Q4_0    |               3.56 |         4 | pp32   |        33.75 |                 35.89 |      1.06 |
|       | llama 7B Q4_0    |               3.56 |         4 | pp64   |        57.97 |                 61.30 |      1.06 |
|       | llama 7B Q4_0    |               3.56 |         4 | pp128  |        89.04 |                 91.20 |      1.02 |
|       | llama 7B Q4_0    |               3.56 |         4 | pp256  |       121.10 |                111.30 |      0.92 |
|       | llama 7B Q4_0    |               3.56 |         4 | pp512  |       128.46 |                117.87 |      0.92 |
|       | llama 7B Q4_0    |               3.56 |        12 | pp32   |        33.56 |                 37.09 |      1.11 |
|       | llama 7B Q4_0    |               3.56 |        12 | pp64   |        57.27 |                 63.48 |      1.11 |
|       | llama 7B Q4_0    |               3.56 |        12 | pp128  |        88.55 |                 97.80 |      1.10 |
|       | llama 7B Q4_0    |               3.56 |        12 | pp256  |       121.35 |                127.76 |      1.05 |
|       | llama 7B Q4_0    |               3.56 |        12 | pp512  |       128.42 |                136.79 |      1.07 |
|       | llama 7B all F32 |              25.10 |         4 | pp32   |        51.04 |                 49.88 |      0.98 |
|       | llama 7B all F32 |              25.10 |         4 | pp64   |        79.66 |                 77.13 |      0.97 |
|       | llama 7B all F32 |              25.10 |         4 | pp128  |       103.58 |                107.29 |      1.04 |
|       | llama 7B all F32 |              25.10 |         4 | pp256  |       141.23 |                123.31 |      0.87 |
|       | llama 7B all F32 |              25.10 |         4 | pp512  |       138.78 |                122.95 |      0.89 |
|       | llama 7B all F32 |              25.10 |        12 | pp32   |        50.99 |                 48.56 |      0.95 |
|       | llama 7B all F32 |              25.10 |        12 | pp64   |        79.38 |                 77.12 |      0.97 |
|       | llama 7B all F32 |              25.10 |        12 | pp128  |       112.81 |                112.74 |      1.00 |
|       | llama 7B all F32 |              25.10 |        12 | pp256  |       141.26 |                140.00 |      0.99 |
|       | llama 7B all F32 |              25.10 |        12 | pp512  |       138.73 |                140.82 |      1.02 |