@ggerganov 2024-06-08
> There is a drop in performance specifically with 4 threads and the larger batch sizes that I don't understand

On M2 Ultra there is a similar effect with the `Q4_0` and `F32` models, while for `Q8_0` and `F16` it behaves as you expect:

```bash
LLAMA_NO_LLAMAFILE=1 LLAMA_NO_METAL=1 ./scripts/compare-commits.sh master sl/blas-backend -m models/tinyllama-1b/ggml-model-q4_0.gguf -m models/tinyllama-1b/ggml-model-q8_0.gguf -m models/tinyllama-1b/ggml-model-f16.gguf -m models/tinyllama-1b/ggml-model-f32.gguf -p 32,64,128,256,512 -n 0 -t 4,8,16
```

| CPU   | Model            |   Model [GiB] |   Threads | Test   |   t/s master |   t/s sl/blas-backend |   Speedup |
|:------|:-----------------|-------------------:|----------:|:-------|-------------:|----------------------:|----------:|
| M2 Ultra | 1B F16     |               2.05 |         4 | pp32   |       265.26 |                214.52 |      0.81 |
| M2 Ultra | 1B F16     |               2.05 |         4 | pp64   |       452.51 |                361.71 |      0.80 |
| M2 Ultra | 1B F16     |               2.05 |         4 | pp128  |       640.73 |                537.69 |      0.84 |
| M2 Ultra | 1B F16     |               2.05 |         4 | pp256  |       664.02 |                588.17 |      0.89 |
| M2 Ultra | 1B F16     |               2.05 |         4 | pp512  |       624.32 |                570.26 |      0.91 |
| M2 Ultra | 1B F16     |               2.05 |         8 | pp32   |       266.72 |                263.57 |      0.99 |
| M2 Ultra | 1B F16     |               2.05 |         8 | pp64   |       464.59 |                441.93 |      0.95 |
| M2 Ultra | 1B F16     |               2.05 |         8 | pp128  |       648.08 |                660.74 |      1.02 |
| M2 Ultra | 1B F16     |               2.05 |         8 | pp256  |       659.51 |                705.26 |      1.07 |
| M2 Ultra | 1B F16     |               2.05 |         8 | pp512  |       622.22 |                745.74 |      1.20 |
| M2 Ultra | 1B F16     |               2.05 |        16 | pp32   |       271.83 |                224.23 |      0.82 |
| M2 Ultra | 1B F16     |               2.05 |        16 | pp64   |       456.11 |                394.74 |      0.87 |
| M2 Ultra | 1B F16     |               2.05 |        16 | pp128  |       633.06 |                610.69 |      0.96 |
| M2 Ultra | 1B F16     |               2.05 |        16 | pp256  |       656.49 |                699.64 |      1.07 |
| M2 Ultra | 1B F16     |               2.05 |        16 | pp512  |       625.12 |                816.75 |      1.31 |
| M2 Ultra | 1B Q4_0    |               0.59 |         4 | pp32   |       227.12 |                224.63 |      0.99 |
| M2 Ultra | 1B Q4_0    |               0.59 |         4 | pp64   |       373.63 |                371.20 |      0.99 |
| M2 Ultra | 1B Q4_0    |               0.59 |         4 | pp128  |       622.08 |                571.99 |      0.92 |
| M2 Ultra | 1B Q4_0    |               0.59 |         4 | pp256  |       658.47 |                612.67 |      0.93 |
| M2 Ultra | 1B Q4_0    |               0.59 |         4 | pp512  |       634.58 |                587.23 |      0.93 |
| M2 Ultra | 1B Q4_0    |               0.59 |         8 | pp32   |       253.38 |                269.27 |      1.06 |
| M2 Ultra | 1B Q4_0    |               0.59 |         8 | pp64   |       410.88 |                436.95 |      1.06 |
| M2 Ultra | 1B Q4_0    |               0.59 |         8 | pp128  |       636.57 |                673.07 |      1.06 |
| M2 Ultra | 1B Q4_0    |               0.59 |         8 | pp256  |       661.17 |                712.17 |      1.08 |
| M2 Ultra | 1B Q4_0    |               0.59 |         8 | pp512  |       631.02 |                745.85 |      1.18 |
| M2 Ultra | 1B Q4_0    |               0.59 |        16 | pp32   |       246.48 |                213.06 |      0.86 |
| M2 Ultra | 1B Q4_0    |               0.59 |        16 | pp64   |       396.73 |                375.92 |      0.95 |
| M2 Ultra | 1B Q4_0    |               0.59 |        16 | pp128  |       632.46 |                600.46 |      0.95 |
| M2 Ultra | 1B Q4_0    |               0.59 |        16 | pp256  |       658.53 |                703.00 |      1.07 |
| M2 Ultra | 1B Q4_0    |               0.59 |        16 | pp512  |       631.18 |                816.15 |      1.29 |
| M2 Ultra | 1B Q8_0    |               1.09 |         4 | pp32   |       248.64 |                221.43 |      0.89 |
| M2 Ultra | 1B Q8_0    |               1.09 |         4 | pp64   |       430.98 |                378.20 |      0.88 |
| M2 Ultra | 1B Q8_0    |               1.09 |         4 | pp128  |       628.65 |                565.11 |      0.90 |
| M2 Ultra | 1B Q8_0    |               1.09 |         4 | pp256  |       650.70 |                606.51 |      0.93 |
| M2 Ultra | 1B Q8_0    |               1.09 |         4 | pp512  |       629.00 |                584.55 |      0.93 |
| M2 Ultra | 1B Q8_0    |               1.09 |         8 | pp32   |       253.84 |                253.78 |      1.00 |
| M2 Ultra | 1B Q8_0    |               1.09 |         8 | pp64   |       434.08 |                431.23 |      0.99 |
| M2 Ultra | 1B Q8_0    |               1.09 |         8 | pp128  |       628.08 |                663.30 |      1.06 |
| M2 Ultra | 1B Q8_0    |               1.09 |         8 | pp256  |       645.79 |                710.54 |      1.10 |
| M2 Ultra | 1B Q8_0    |               1.09 |         8 | pp512  |       629.27 |                753.74 |      1.20 |
| M2 Ultra | 1B Q8_0    |               1.09 |        16 | pp32   |       252.36 |                214.02 |      0.85 |
| M2 Ultra | 1B Q8_0    |               1.09 |        16 | pp64   |       431.50 |                372.52 |      0.86 |
| M2 Ultra | 1B Q8_0    |               1.09 |        16 | pp128  |       635.57 |                589.64 |      0.93 |
| M2 Ultra | 1B Q8_0    |               1.09 |        16 | pp256  |       644.00 |                707.12 |      1.10 |
| M2 Ultra | 1B Q8_0    |               1.09 |        16 | pp512  |       634.01 |                813.76 |      1.28 |
| M2 Ultra | 1B all F32 |               4.10 |         4 | pp32   |       324.34 |                318.84 |      0.98 |
| M2 Ultra | 1B all F32 |               4.10 |         4 | pp64   |       497.14 |                501.49 |      1.01 |
| M2 Ultra | 1B all F32 |               4.10 |         4 | pp128  |       780.47 |                687.67 |      0.88 |
| M2 Ultra | 1B all F32 |               4.10 |         4 | pp256  |       719.40 |                676.08 |      0.94 |
| M2 Ultra | 1B all F32 |               4.10 |         4 | pp512  |       668.30 |                614.47 |      0.92 |
| M2 Ultra | 1B all F32 |               4.10 |         8 | pp32   |       394.99 |                325.57 |      0.82 |
| M2 Ultra | 1B all F32 |               4.10 |         8 | pp64   |       588.41 |                522.05 |      0.89 |
| M2 Ultra | 1B all F32 |               4.10 |         8 | pp128  |       769.56 |                727.87 |      0.95 |
| M2 Ultra | 1B all F32 |               4.10 |         8 | pp256  |       715.89 |                742.46 |      1.04 |
| M2 Ultra | 1B all F32 |               4.10 |         8 | pp512  |       672.34 |                772.63 |      1.15 |
| M2 Ultra | 1B all F32 |               4.10 |        16 | pp32   |       397.21 |                246.50 |      0.62 |
| M2 Ultra | 1B all F32 |               4.10 |        16 | pp64   |       589.79 |                421.42 |      0.71 |
| M2 Ultra | 1B all F32 |               4.10 |        16 | pp128  |       770.60 |                655.37 |      0.85 |
| M2 Ultra | 1B all F32 |               4.10 |        16 | pp256  |       712.96 |                724.44 |      1.02 |
| M2 Ultra | 1B all F32 |               4.10 |        16 | pp512  |       669.12 |                837.94 |      1.25 |