@slaren 2024-06-08
I realized that there is an issue that causes the `kqv` matrix multiplication to not be offloaded to the BLAS backend, and this is what causes the performance difference with a low number of threads. If this matrix multiplication is removed by enabling flash attention, then the performance difference disappears.
| CPU   | Model            |   Model Size [GiB] |   Threads | Test   |   t/s master |   t/s sl/blas-backend |   Speedup |
|:------|:-----------------|-------------------:|----------:|:-------|-------------:|----------------------:|----------:|
| M3 Max | 7B Q4_0    |               3.56 |         4 | pp32   |        32.77 |                 36.23 |      1.11 |
| M3 Max | 7B Q4_0    |               3.56 |         4 | pp64   |        58.37 |                 62.11 |      1.06 |
| M3 Max | 7B Q4_0    |               3.56 |         4 | pp128  |        92.00 |                 95.05 |      1.03 |
| M3 Max | 7B Q4_0    |               3.56 |         4 | pp256  |       119.94 |                121.50 |      1.01 |
| M3 Max | 7B Q4_0    |               3.56 |         4 | pp512  |       127.86 |                126.98 |      0.99 |
| M3 Max | 7B Q4_0    |               3.56 |        12 | pp32   |        33.67 |                 38.43 |      1.14 |
| M3 Max | 7B Q4_0    |               3.56 |        12 | pp64   |        57.92 |                 64.87 |      1.12 |
| M3 Max | 7B Q4_0    |               3.56 |        12 | pp128  |        90.62 |                 99.64 |      1.10 |
| M3 Max | 7B Q4_0    |               3.56 |        12 | pp256  |       120.33 |                129.79 |      1.08 |
| M3 Max | 7B Q4_0    |               3.56 |        12 | pp512  |       126.31 |                141.03 |      1.12 |
| M3 Max | 7B F32 |              25.10 |         4 | pp32   |        51.18 |                 50.64 |      0.99 |
| M3 Max | 7B F32 |              25.10 |         4 | pp64   |        79.80 |                 79.86 |      1.00 |
| M3 Max | 7B F32 |              25.10 |         4 | pp128  |       112.13 |                111.49 |      0.99 |
| M3 Max | 7B F32 |              25.10 |         4 | pp256  |       138.19 |                133.77 |      0.97 |
| M3 Max | 7B F32 |              25.10 |         4 | pp512  |       133.84 |                131.51 |      0.98 |
| M3 Max | 7B F32 |              25.10 |        12 | pp32   |        50.47 |                 48.97 |      0.97 |
| M3 Max | 7B F32 |              25.10 |        12 | pp64   |        79.65 |                 78.19 |      0.98 |
| M3 Max | 7B F32 |              25.10 |        12 | pp128  |       112.97 |                115.07 |      1.02 |
| M3 Max | 7B F32 |              25.10 |        12 | pp256  |       137.53 |                143.82 |      1.05 |
| M3 Max | 7B F32 |              25.10 |        12 | pp512  |       134.25 |                149.98 |      1.12 |